Convert ddp-client to use only api.mainModule for client and server
