+++ packages/ddp-client/client.js
@@ -1,0 +1,14 @@
+import { DDP, LivedataTest } from './namespace';
+
+import './livedata_common';
+
+import './sockjs-0.3.4';
+import './stream_client_sockjs';
+import './stream_client_common';
+
+import './random_stream';
+import './livedata_connection';
+
+import './client_convenience';
+
+export { DDP, LivedataTest };
