+++ packages/ddp-client/package.js
@@ -27,19 +27,10 @@ Package.onUse(function (api) {
   // _idParse, _idStringify.
   api.use('mongo-id', ['client', 'server']);
 
-  api.addFiles(['sockjs-0.3.4.js', 'stream_client_sockjs.js'], 'client');
-  api.addFiles('stream_client_nodejs.js', 'server');
-  api.addFiles('stream_client_common.js', ['client', 'server']);
-
-  api.addFiles('livedata_common.js', ['client', 'server']);
-  api.addFiles('random_stream.js', ['client', 'server']);
-
-  api.addFiles('livedata_connection.js', ['client', 'server']);
-
-  api.addFiles('client_convenience.js', 'client');
-
-  api.mainModule("namespace.js");
+  // For backcompat where things use Package.ddp.DDP, etc
   api.export('DDP');
+  api.mainModule("client.js", "client");
+  api.mainModule("server.js", "server");
 });
 
 Package.onTest(function (api) {
