+++ packages/ddp-client/server.js
@@ -1,0 +1,11 @@
+import { DDP, LivedataTest } from './namespace';
+
+import './livedata_common';
+
+import './stream_client_nodejs';
+import './stream_client_common';
+
+import './random_stream';
+import './livedata_connection';
+
+export { DDP, LivedataTest };
